Refactor VideoDetailViewController.
