+++ SwiftVideos/Cartfile.resolved
@@ -4,9 +4,9 @@ github "RxSwiftCommunity/Action" "4.0.1"
 github "RxSwiftCommunity/RxDataSources" "4.0.1"
 github "SnapKit/SnapKit" "5.0.1"
 github "cbpowell/MarqueeLabel" "4.0.0"
+github "josejuanqm/VersaPlayer" "3.0.0"
 github "libgit2/objective-git" "0.14.2"
 github "mxcl/PromiseKit" "6.10.0"
 github "onevcat/Kingfisher" "5.7.1"
-github "piemonte/Player" "0.13.1"
 github "quickbirdstudios/XCoordinator" "0963e9faf25b6015da9d1fb395c107b6d840d4e4"
 github "rinov/YoutubeKit" "3a665e6659252599b9e58a1596566bae67bdd671"
